refactor: removed integer division

+++ UnitTestingDemo/Mathematics.cs
@@ -7,7 +7,7 @@
             return x * y;
         }
 
-        public decimal Divide(int x, int y)
+        public decimal Divide(decimal x, decimal y)
         {
             return x / y;
         }
